requirements.txt
klein
twisted
service_identity
protobuf==3.20.1
git+https://github.com/spotware/OpenApiPy.git
python-dotenv
loguru
